+++ packages/studio-explore/src/components/Connection/index.tsx
@@ -1,17 +1,19 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useContext } from '@graphscope/studio-graph';
-import { SplitSection } from '@graphscope/studio-components';
-import { theme, Flex, Input, Modal, Button } from 'antd';
-import { DatabaseOutlined } from '@ant-design/icons';
-import LoadKuzuWasm from './import-from-csv';
+import { SplitSection, Utils } from '@graphscope/studio-components';
+import { theme, Flex, Input, Modal, Button, Typography } from 'antd';
+import { ApiOutlined } from '@ant-design/icons';
+import ImportIntoKuzu from './import-into-kuzu';
 import { ConnectEndpoint } from '@graphscope/studio-query';
 import { LoadCSV } from '@graphscope/studio-graph';
+
 interface IInitProps {}
 
 const Init: React.FunctionComponent<IInitProps> = props => {
   const { id, store } = useContext();
   const { token } = theme.useToken();
   const [visible, setVisible] = React.useState(false);
+
   const handleToggle = () => {
     setVisible(!visible);
   };
@@ -19,14 +21,29 @@ const Init: React.FunctionComponent<IInitProps> = props => {
     setVisible(false);
   };
   const handleConnect = params => {
-    console.log('params');
     handleClose();
   };
+  useEffect(() => {
+    const query_endpoint = Utils.storage.get('query_endpoint');
+    if (!query_endpoint) {
+      setVisible(true);
+    }
+  }, []);
   return (
     <>
-      <Button onClick={handleToggle} icon={<DatabaseOutlined />} type="text"></Button>
+      <Button onClick={handleToggle} icon={<ApiOutlined />} type="text"></Button>
       <Modal open={visible} onClose={handleClose} onCancel={handleClose} width={'80%'} footer={null}>
-        <SplitSection leftSide={<ConnectEndpoint onConnect={handleConnect} />} rightSide={<LoadCSV />} />
+        <SplitSection
+          leftSide={<ConnectEndpoint onConnect={handleConnect} />}
+          rightSide={
+            <Flex vertical style={{ padding: '12px', display: 'flex', flex: 1, height: '90%' }}>
+              <Typography.Title level={3} style={{ marginBottom: '16px' }}>
+                Load CSV Files
+              </Typography.Title>
+              <ImportIntoKuzu handleClose={handleClose} />
+            </Flex>
+          }
+        />
       </Modal>
     </>
   );