+++ packages/studio-graph/src/components/SwitchEngine/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Button, TooltipProps, Tooltip } from 'antd';
+import { Button, TooltipProps, Tooltip, theme } from 'antd';
 import { FormattedMessage } from 'react-intl';
 import { useContext } from '../../hooks/useContext';
 import { Icons, useStudioProvier } from '@graphscope/studio-components';
@@ -12,15 +12,20 @@ export interface ISwitchEngineProps {
 const SwitchEngine: React.FunctionComponent<ISwitchEngineProps> = props => {
   const { updateStore, store } = useContext();
   const { render } = store;
+  const { token } = theme.useToken();
   const { title = 'Switch graph view dimensions', placement = 'left' } = props;
-  const { isLight } = useStudioProvier();
-  const color = !isLight ? '#ddd' : '#000';
+
   const handleSwitch = () => {
     updateStore(draft => {
       draft.render = render === '2D' ? '3D' : '2D';
     });
   };
-  const icon = render === '2D' ? <Graph3D style={{ color }} /> : <Graph2D style={{ color }} />;
+  const icon =
+    render === '2D' ? (
+      <Graph3D style={{ color: token.colorTextBase }} />
+    ) : (
+      <Graph2D style={{ color: token.colorTextBase }} />
+    );
   return (
     <Tooltip title={<FormattedMessage id={`${title}`} />} placement="left">
       <Button onClick={handleSwitch} icon={icon} type="text"></Button>