+++ packages/studio-website/src/global.d.ts
@@ -1,3 +1,4 @@
 interface Window {
   GS_ENGINE_TYPE: 'groot' | 'interactive' | 'gart';
+  COORDINATOR_URL: string;
 }