+++ packages/studio-explore/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphscope/studio-explore
 
+## 0.1.6
+
+### Patch Changes
+
+- update site
+- Updated dependencies
+  - @graphscope/studio-components@0.1.8
+  - @graphscope/studio-driver@0.1.7
+  - @graphscope/studio-graph@0.1.8
+  - @graphscope/studio-importor@0.1.8
+  - @graphscope/studio-query@0.1.8
+
 ## 0.1.5
 
 ### Patch Changes