+++ packages/studio-graph-editor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphscope/studtio-graph-editor
 
+## 0.1.2
+
+### Patch Changes
+
+- update site
+- Updated dependencies
+  - @graphscope/studio-components@0.1.8
+
 ## 0.1.1
 
 ### Patch Changes